Remove token from api calls, move solely into socket
