+++ assets/src/hooks/useRoutes.ts
@@ -1,11 +1,11 @@
 import { useEffect, useState } from "react"
 import { fetchRoutes } from "../api"
-import { Route, UserToken } from "../skate"
+import { Route } from "../skate"
 
-const useRoutes = (userToken?: UserToken): Route[] | null => {
+const useRoutes = (): Route[] | null => {
   const [routes, setRoutes] = useState<Route[] | null>(null)
   useEffect(() => {
-    fetchRoutes(userToken).then((newRoutes: Route[]) => setRoutes(newRoutes))
+    fetchRoutes().then((newRoutes: Route[]) => setRoutes(newRoutes))
   }, [])
   return routes
 }
