+++ assets/src/hooks/useSocket.ts
@@ -2,8 +2,20 @@ import { Socket } from "phoenix"
 import { useEffect, useState } from "react"
 import { UserToken } from "../skate"
 
-const useSocket = (userToken?: UserToken): Socket | undefined => {
+export const readUserToken = (): UserToken | undefined => {
+  const dataEl = document.getElementById("app")
+  if (!dataEl) {
+    return undefined
+  }
+
+  const token = dataEl.dataset.userToken as UserToken
+  return token
+}
+
+const useSocket = (): Socket | undefined => {
   const [socket, setSocket] = useState<Socket | undefined>(undefined)
+
+  const userToken: UserToken | undefined = readUserToken()
 
   useEffect(() => {
     const initialSocket = new Socket("/socket", {
